fix!: Move `Margin` to `renderer`
